And class hierarchy to Entry
